delete wishlist

+++ processes/process-delete-wishlist.php
@@ -6,20 +6,24 @@ include '../includes/db-connect.php';
 $listing_id = $_GET['del_wish'];
 $user_id = $_SESSION['user_id'];
 
-if (isset($_SESSION['guestUserName'])) {
 
-    $stmt = $mysqli->prepare("DELETE FROM wishlist WHERE listing_id = ?");
-$stmt->bind_param("i", $listing_id);
-     $stmt->execute();
+
+    $stmt = $mysqli->prepare("DELETE FROM wishlist WHERE listing_id = ? AND user_id = ?");
+$stmt->bind_param("ii", $listing_id, $user_id);
+     
  
  if (!$stmt->execute()) {
-    $_SESSION["listDelError"] = "<div class='error-message'>Oops! Something went wrong... (" . $stmt->errno . ") " . $stmt->error. ".<br>Please see your website administrator and quote this error message.</div>";
+    $_SESSION["wlError"] = "<div class='error-message'>Oops! Something went wrong... (" . $stmt->errno . ") " . $stmt->error. ".<br>Please see your website administrator and quote this error message.</div>";
+    header('location: ../wishlist');
+    exit;
 } else {
-    $_SESSION["successMessage"] = "<div class='success-message'>User ID: " . $listing_id ." was successfully deleted</div>";
+     $_SESSION['wlSuccess'] = "Listing with ID: " . $listing_id . " was successfully removed from your wishlist ";
+        header('location: ../wishlist');
 }
-}
+
+
 $stmt->close();
 
 
 
-// Remove from array
+
